refactor: item 컴포넌트 분리
