+++ src/features/side/index.tsx
@@ -1,11 +1,10 @@
 import { useState } from "react";
 import { useQuery } from "@tanstack/react-query";
-import { NavLink, useNavigate } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 import { RxHamburgerMenu } from "react-icons/rx";
 import {
     GrFormPrevious,
     GrFormNext,
-    GrFormClose,
     GrFormAdd,
 } from "react-icons/gr";
 import {
@@ -26,12 +25,12 @@ import {
 import { usePagination } from "./hook/usePagenation";
 import { fetchSidebarData } from "./api/sideApi";
 import { DeleteConfirm } from "./ui/deleteConfirm";
+import { Item } from "./ui/Item";
 
 const Sidebar = () => {
     const [idToDelete, setIdToDelete] = useState<string | null>(null);
     const [isDialogOpen, setDialogOpen] = useState(false);
 
-    const navigate = useNavigate();
     const { data, isError, error } = useQuery({
         queryKey: ["side"],
         queryFn: fetchSidebarData,
@@ -78,29 +77,10 @@ const Sidebar = () => {
                 <PopoverBody>
                     <Box minHeight="400px">
                         {paginatedItems.map((item) => (
-                            <Flex
-                                key={item.id}
-                                align="center"
-                                p="2"
-                                rounded="md"
-                                title={item.label}
-                                _hover={{ bg: "gray.100" }}
-                                cursor="pointer"
-                                onClick={() => navigate(`/${item.id}`)}
-                            >
-                                <Text ml="4" minWidth="200px">
-                                    {item.label}
-                                </Text>
-                                <Button
-                                    ml="auto"
-                                    bg="none"
-                                    color="gray"
-                                    cursor="pointer"
-                                    onClick={() => handleDelete(item.id)}
-                                >
-                                    <GrFormClose />
-                                </Button>
-                            </Flex>
+                            <Item
+                                item={item}
+                                handleDelete={handleDelete}
+                            />
                         ))}
                         {isError && (
                             <Text color="red.500">{error.message}</Text>
